Fix compiler warning
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@21865 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/algo/gnomon/gnomon.cpp
@@ -299,7 +299,7 @@ void CGnomon::Run(void)
         vector<IPair> mrna_vec;
         vector<IPair> cds_vec;
 
-        for (int j = 0;  j < igene.size();  ++j) {
+        for (size_t j = 0;  j < igene.size();  ++j) {
             const ExonData& exon = igene[j];
             int a = exon.Start();
             int b = exon.Stop();
@@ -393,6 +393,9 @@ END_NCBI_SCOPE
 /*
  * ===========================================================================
  * $Log$
+ * Revision 1.10  2005/04/26 17:28:26  dicuccio
+ * Fix compiler warning
+ *
  * Revision 1.9  2004/11/18 21:27:40  grichenk
  * Removed default value for scope argument in seq-loc related functions.
  *
